+++ packages/sdk/src/lib/dataProtectorSharing/IExecDataProtectorSharing.ts
@@ -20,8 +20,6 @@ import {
   GetProtectedDataPricingParamsResponse,
   GetRentalsParams,
   GetRentalsResponse,
-  GetResultFromCompletedTaskParams,
-  GetResultFromCompletedTaskResponse,
   GetUserAppWhitelistParams,
   GetUserAppWhitelistResponse,
   RemoveCollectionParams,
@@ -50,7 +48,6 @@ import { getCollectionSubscriptions } from './getCollectionSubscriptions.js';
 import { getProtectedDataInCollections } from './getProtectedDataInCollections.js';
 import { getProtectedDataPricingParams } from './getProtectedDataPricingParams.js';
 import { getRentals } from './getRentals.js';
-import { getResultFromCompletedTask } from './getResultFromCompletedTask.js';
 import { getUserAddOnlyAppWhitelist } from './getUserAddOnlyAppWhitelist.js';
 import { removeCollection } from './removeCollection.js';
 import { removeProtectedDataForSale } from './removeProtectedDataForSale.js';
@@ -251,16 +248,6 @@ class IExecDataProtectorSharing extends IExecDataProtectorModule {
       ...args,
       iexec: this.iexec,
       sharingContractAddress: this.sharingContractAddress,
-    });
-  }
-
-  async getResultFromCompletedTask(
-    args: GetResultFromCompletedTaskParams
-  ): Promise<GetResultFromCompletedTaskResponse> {
-    await isValidProvider(this.iexec);
-    return getResultFromCompletedTask({
-      ...args,
-      iexec: this.iexec,
     });
   }