+++ packages/sdk/CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
-- Move move GetResultFromCompletedTask from DPS to DPC
+- Move move `GetResultFromCompletedTask` from DPS to DPC
 
 ## [2.0.0-beta.11] (2025-01-29)