+++ packages/sdk/CHANGELOG.md
@@ -6,7 +6,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
-Moved `getResultFromCompletedTask` method from DataProtector Sharing module (DPS) to DataProtector Core module (DPC).
+- Moved `getResultFromCompletedTask` method from DataProtectorSharing module (DPS) to DataProtectorCore module (DPC).
+
 ## [2.0.0-beta.11] (2025-01-29)
 
 ### Added